third_party/blink/web_tests/external/wpt/css/css-viewport/zoom/svg-computed-style-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-underline-offset value '4px' no zoom
  assert_equals: expected "4px" but got "auto"
[FAIL] Property text-underline-offset value 'inherit' no zoom
  assert_equals: expected "2px" but got "auto"
[FAIL] Property text-underline-offset value '4px' zoom: 2
  assert_equals: expected "4px" but got "auto"
[FAIL] Property text-underline-offset value 'inherit' zoom: 2
  assert_equals: expected "2px" but got "auto"
Harness: the test ran to completion.
